Fix names with cyrillic characters and make it a sync
