+++ composer.json
@@ -70,7 +70,8 @@
         "spipu/html2pdf": "^5.2",
         "vinkla/hashids": "^12.0",
         "willvincent/feeds": "^2.4",
-        "setasign/tfpdf": "^1.33"
+        "setasign/tfpdf": "^1.33",
+        "ext-iconv": "*"
     },
     "require-dev": {
         "barryvdh/laravel-ide-helper": "^3.0",
